requirements.txt
alembic==1.4.2
attrs==19.3.0
Automat==20.2.0
beautifulsoup4==4.9.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.1
Mako==1.1.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
parsel==1.6.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
queuelib==1.5.0
requests==2.23.0
Scrapy==2.1.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.17
Twisted==20.3.0
urllib3==1.25.9
w3lib==1.22.0
Werkzeug==1.0.1
zope.interface==5.1.0
